+++ classes/MyCMS.php
@@ -2,6 +2,8 @@
 
 namespace WorkOfStan\MyCMS;
 
+use WorkOfStan\MyCMS\L10n;
+
 /**
  * Extension of a MyCMS object with translations.
  * It holds all variables needed for the used project.
@@ -17,18 +19,18 @@ class MyCMS extends MyCMSMonoLingual
     use \Nette\SmartObject;
 
     /**
-     * Selected locale strings
+     * Folder and name prefix of localisation yml
      *
-     * @var array<string>
+     * @var L10n
      */
-    public $TRANSLATION;
+    protected $localisation;
 
     /**
-     * Available languages
+     * Folder and name prefix of localisation yml
      *
-     * @var array<string>
+     * @var string
      */
-    public $TRANSLATIONS;
+    protected $prefixL10n;
 
     /**
      * PARAMETRIC URL into TEMPLATE conditions (for FriendlyURL functionality)
@@ -45,6 +47,13 @@ class MyCMS extends MyCMSMonoLingual
     public $typeToTableMapping;
 
     /**
+     * Available languages
+     *
+     * @var array<string>
+     */
+    public $TRANSLATIONS;
+
+    /**
      * Constructor
      *
      * @param array<mixed> $myCmsConf
@@ -52,6 +61,7 @@ class MyCMS extends MyCMSMonoLingual
     public function __construct(array $myCmsConf = [])
     {
         parent::__construct($myCmsConf);
+        $this->localisation = new L10n($this->prefixL10n, $this->TRANSLATIONS);
     }
 
     /**
@@ -77,19 +87,7 @@ class MyCMS extends MyCMSMonoLingual
                 ) ? $sessionArray['language'] : DEFAULT_LANGUAGE
             );
         if ($makeInclude) {
-            $languageFile = DIR_TEMPLATE . '/../language-' . $resultLanguage . '.inc.php';
-            if (!file_exists($languageFile)) {
-                throw new \Exception("Missing expected language file {$languageFile}");
-            }
-            // include (as include_once triggers error in PHPUnit tests because of attempted repeated includes)
-            include $languageFile; // MUST contain $translation = [...];
-            // TODO instead of PHP file with array $translation, let's put localized texts into yaml.
-            /** @phpstan-ignore-next-line */
-            if (!(isset($translation) && is_array($translation))) {
-                throw new \Exception("Missing expected translation {$languageFile}");
-            }
-            /** @phpstan-ignore-next-line */
-            $this->TRANSLATION = $translation;
+            $this->localisation->loadLocalisation($resultLanguage);
         }
         return $resultLanguage;
     }
@@ -97,6 +95,7 @@ class MyCMS extends MyCMSMonoLingual
     /**
      * Translate defined string to the language stored in $_SESSION['language'].
      * Returns original text if translation not found.
+     * TODO: refactor this wrapper away
      *
      * @param string $id text to translate
      * @param int|null $options case transposition - null || [MB_CASE_UPPER|MB_CASE_LOWER|MB_CASE_TITLE|L_UCFIRST]
@@ -104,20 +103,6 @@ class MyCMS extends MyCMSMonoLingual
      */
     public function translate($id, $options = null)
     {
-        if (
-            !isset($this->TRANSLATION[$id]) && isset($_SESSION['test-translations'])
-            && $_SESSION['language'] != DEFAULT_LANGUAGE
-        ) {
-            $this->logger->warning('Translation does not exist - ' . $id);
-        }
-        $result = isset($this->TRANSLATION[$id]) ? $this->TRANSLATION[$id] : $id;
-        if ($options === L_UCFIRST) {
-            $result = mb_strtoupper(mb_substr($result, 0, 1)) . mb_substr($result, 1);
-        } elseif (
-            is_int($options) && ($options == MB_CASE_UPPER || $options == MB_CASE_LOWER || $options == MB_CASE_TITLE)
-        ) {
-            $result = mb_convert_case($result, $options);
-        }
-        return $result;
+        return $this->localisation->translate($id, $options);
     }
 }